pyproject.toml
[tool.poetry]
name = "stable-diffusion-prompt-generator"
version = "0.1.0"
description = "A Python tool that uses Claude to generate optimized prompts for Stability AI's image generation models"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "sdprompt", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
anthropic = "^0.18.1"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0"
click = "^8.1.7"
rich = "^13.7.0"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sdprompt = "sdprompt.main:main" 